% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Estimate latent class models with different number of classes
this function takes the maximum number of classes one wants to run
then runs models from 1 class to the maximum amount of classes
outputs the result as a list
user needs to define the model, maximum iterations and number of repetitions
Estimate required LC models with different number of classes}
\usage{
models(
  formula,
  data,
  k,
  maxiter,
  graphs,
  tol,
  na.rm,
  probs.start,
  nrep,
  verbose,
  calc.se
)
}
\arguments{
\item{formula}{The latent class model (as documented in poLCA)}

\item{data}{The dataset used}

\item{k}{The maximum number of classes required}

\item{maxiter}{The maximum number of iterations}

\item{graphs}{Shall the output include graphs?}

\item{tol}{Tolerance}

\item{na.rm}{Remove missing values?}

\item{probs.start}{If starting values are to be entered manually}

\item{nrep}{Number of times that model should be estimated using different starting values}

\item{verbose}{NOT SURE CHECK}

\item{calc.se}{Calculate standard errors?}
}
\value{
A list
}
\description{
Estimate latent class models with different number of classes
this function takes the maximum number of classes one wants to run
then runs models from 1 class to the maximum amount of classes
outputs the result as a list
user needs to define the model, maximum iterations and number of repetitions
Estimate required LC models with different number of classes
}
\examples{
Estimate latent class models with 1,2,3 and 4 classes
Each model to iterate up to 5,000 times and
repeated 10 times.
models(data, 4, f, 5000, 10)
}
