% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateModels.R
\name{estimateModels}
\alias{estimateModels}
\title{Estimate latent class models with different numbers of classes}
\usage{
estimateModels(
  formula,
  data,
  k,
  maxiter,
  graphs,
  tol,
  na.rm,
  probs.start,
  nrep,
  verbose,
  calc.se
)
}
\arguments{
\item{formula}{The latent class model (as documented in the poLCA package).}

\item{data}{A dataframe as documented in the poLCA package.}

\item{k}{The maximum number of classes required.}

\item{...}{All other poLCA() arguments as documented in the poLCA package.}
}
\value{
estimateModels() returns a list of poLCA objects (that are also lists).
Each element of the list is the poLCA object for each the latent class
models estimated (i.e. the parameter estimates and other model information
of the latent class models with one, two, etc classes). Please see package
poLCA for full documentation.
}
\description{
Use this function to estimate latent class models with different
numbers of classes. Results for all models are given in a
single object.
}
\details{
This is a convenience function, which uses the function
\code{\link{poLCA}} from the \code{\link[poLCA]} package
to estimate latent class models with a user defined
maximum number of classes (k). The function will estimate
latent class models of 1:k classes. The output is a list,
each element of which is the poLCA() output object
(also a list) for each of the latent class models estimated.
}
\examples{
Estimate latent class models with 1,2,3 and 4 classes
Each model to iterate up to 5,000 times and repeated 10 times.
data(election)
f <- cbind(MORALG,CARESG,KNOWG,LEADG,DISHONG,INTELG,MORALB,CARESB,KNOWB,LEADB,DISHONB,INTELB) ~ 1
lc_models <- estimateModels(formula = f, data = election, k = 4, maxiter = 5000, nrep = 10)
}
